+++ .env
@@ -28,12 +28,12 @@ MF_JAEGER_URL=jaeger:6831
 MF_AUTH_LOG_LEVEL=debug
 MF_AUTH_HTTP_PORT=8189
 MF_AUTH_GRPC_PORT=8181
-MF_AUTH_GRPC_URL=authn:8181
+MF_AUTH_GRPC_URL=auth:8181
 MF_AUTH_GRPC_TIMEOUT=1s
 MF_AUTH_DB_PORT=5432
 MF_AUTH_DB_USER=mainflux
 MF_AUTH_DB_PASS=mainflux
-MF_AUTH_DB=authn
+MF_AUTH_DB=auth
 MF_AUTH_SECRET=secret
 
 ### Users
